well now it kinda works with 500k starting tree, i have never seen struct greater in practice, so i will stop. yes it is still not optimized, but it was fun. and also it is fun in how and why now it works

+++ structuresOfAllKind/structuresOfAllKind/MyBinaryTree.cpp
@@ -34,7 +34,7 @@ BinaryTreeNode* MyBinaryTree::find(long long key)
 	return root.key == key? &root: findNextStep(&root, key);
 }
 
-bool MyBinaryTree::insert(int key, int value)
+bool MyBinaryTree::insert(long long key, long long value)
 {
 	BinaryTreeNode* node;
 	if ((node = find(value)) != nullptr)
